Finished installer, working ok

+++ uninstall.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
+if [ "$EUID" -ne 0 ]
+  then echo "Please run as root"
+  exit
+fi
 rm -rf /usr/local/share/quantum_chess
-rm build/src/server/server /usr/local/bin/quantum_chess_server
-rm build/src/client/client /usr/local/bin/quantum_chess
-# REMOVE DESKTOP
+rm /usr/local/bin/quantum_chess_server
+rm /usr/local/bin/quantum_chess
+rm /usr/local/bin/__quantum_chess_server
+rm /usr/local/bin/__quantum_chess
+rm /usr/share/applications/quantum_chess.desktop
